add cache.ts into common
