+++ packages/common/src/index.ts
@@ -2,4 +2,5 @@ export { default as Segment, validateSegments } from "./segment";
 export { default as Message } from "./message";
 export { default as Request } from "./request";
 export { default as Response } from "./response";
+export { default as Cache } from "./cache";
 export * as utils from "./utils";
