+++ packages/sdk/src/index.ts
@@ -1,3 +1,21 @@
-import { Manager } from "./manager";
+import { Cache, Request, Response } from "rpch-commons";
+import RequestCache from "./request-cache";
 
-console.log("we got a manager!", !!Manager);
+// export default class SDK {
+//   private cache: Cache
+//   private requestCache: RequestCache
+
+//   constructor() {
+//     this.cache = new Cache
+//   }
+
+//   public createRequest(
+//     origin: string,
+//     provider: string,
+//     destination: string
+//   ): Request {
+//     return Request.fromData(origin, provider, destination);
+//   }
+
+//   public sendRequest(req: Request): Promise<Response> {}
+// }
